requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
openai==1.61.0
scikit-learn==1.6.1
# hdbscan==0.8.40
numpy==2.2.2
python-pptx==1.0.2
gunicorn==23.0.0
python-dotenv==1.0.1
pydantic==2.10.5
whitenoise==6.8.2
